// ReweReceipt.Web/Client/src/api/schemas/articleReceipt.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * ReweReceipt.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ArticleReceiptQuantity } from "./articleReceiptQuantity";
import type { ArticleReceiptPrice } from "./articleReceiptPrice";

export interface ArticleReceipt {
  receiptId: string;
  timeStamp: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  quantity: ArticleReceiptQuantity;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  price: ArticleReceiptPrice;
}